+++ android/src/facebook/LikeButtonView.java
@@ -1,9 +1,14 @@
 /**
-  * Copyright (c) 2014 by Mark Mokryn All Rights Reserved.
-  * Licensed under the terms of the Apache Public License 2.0
-  * Please see the LICENSE included with this distribution for details.
-  *
-  */
+ * Copyright (c) 2014 by Mark Mokryn All Rights Reserved.
+ * Licensed under the terms of the Apache Public License 2.0
+ * Please see the LICENSE included with this distribution for details.
+ *
+ * Appcelerator Titanium Mobile
+ * Copyright (c) 2015 by Appcelerator, Inc. All Rights Reserved.
+ * Licensed under the terms of the Apache Public License
+ * Please see the LICENSE included with this distribution for details.
+ *
+ */
 
 package facebook;
 
@@ -15,21 +20,21 @@ import org.appcelerator.kroll.common.Log;
 import org.appcelerator.titanium.util.TiColorHelper;
 import org.appcelerator.titanium.util.TiConvert;
 
-import com.facebook.widget.LikeView.*;
-import com.facebook.widget.LikeView;
+import com.facebook.share.widget.LikeView;
+import com.facebook.share.widget.LikeView.AuxiliaryViewPosition;
+import com.facebook.share.widget.LikeView.HorizontalAlignment;
+import com.facebook.share.widget.LikeView.Style;
+
 
 public class LikeButtonView extends TiUIView 
 {
-	// Standard Debugging variables
 	private static final String TAG = "LikeButtonView";
 	private LikeView likeView;
-	
+		
 	public LikeButtonView(TiViewProxy proxy) 
 	{
 		super(proxy);
-		
 		Log.d(TAG, "[VIEW LIFECYCLE EVENT] view");
-		
 		likeView = new LikeView(proxy.getActivity());
 		// Set the view as the native view. You must set the native view
 		// for your view to be rendered correctly.
@@ -40,10 +45,9 @@ public class LikeButtonView extends TiUIView
 	// is realized by the view proxy. That means that the processProperties
 	// method will be called during creation and that propertiesChanged and 
 	// propertyChanged will be called when properties are changed on the proxy.
-	
 	private void setLikeViewStyle(String name) {
 		for(Style style : Style.values()) { 
-			if (style.toString().equalsIgnoreCase(name)) {
+			if (style.name().toString().equalsIgnoreCase(name)) {
 				likeView.setLikeViewStyle(style);
 				break;
 			}
@@ -52,7 +56,7 @@ public class LikeButtonView extends TiUIView
 
 	private void setAuxiliaryViewPosition(String name) {
 		for(AuxiliaryViewPosition avp : AuxiliaryViewPosition.values()) { 
-			if (avp.toString().equalsIgnoreCase(name)) {
+			if (avp.name().toString().equalsIgnoreCase(name)) {
 				likeView.setAuxiliaryViewPosition(avp);
 				break;
 			}
@@ -61,7 +65,7 @@ public class LikeButtonView extends TiUIView
 	
 	private void setHorizontalAlignment(String name) {
 		for(HorizontalAlignment ha : HorizontalAlignment.values()) { 
-			if (ha.toString().equalsIgnoreCase(name)) {
+			if (ha.name().toString().equalsIgnoreCase(name)) {
 				likeView.setHorizontalAlignment(ha);
 				break;
 			}
@@ -74,10 +78,9 @@ public class LikeButtonView extends TiUIView
 		super.processProperties(props);
 		Log.d(TAG,"[VIEW LIFECYCLE EVENT] processProperties " + props);
 		if (props.containsKey("objectId")) {
-			likeView.setObjectId(props.getString("objectId"));
+			likeView.setObjectIdAndType(props.getString("objectID"),LikeView.ObjectType.PAGE);
 		}
 		if (props.containsKey("likeViewStyle")) {
-			//likeView.setLikeViewStyle(Style.values()[props.getInt("likeViewStyle")]);
 			String styleName = props.getString("likeViewStyle");
 			setLikeViewStyle(styleName);
 
@@ -100,9 +103,8 @@ public class LikeButtonView extends TiUIView
 	{
 		// This method is called whenever a proxy property value is updated. Note that this 
 		// method is only called if the new value is different than the current value.
-
-		if (key.equals("objectId")) {
-			likeView.setObjectId(TiConvert.toString(newValue));
+		if (key.equals("objectID")) {
+			likeView.setObjectIdAndType(TiConvert.toString(newValue),LikeView.ObjectType.PAGE);
 		} else if (key.equals("likeViewStyle")) {
 			setLikeViewStyle(TiConvert.toString(newValue));
 		} else if (key.equals("auxiliaryViewPosition")) {
@@ -116,4 +118,5 @@ public class LikeButtonView extends TiUIView
 		}
 		Log.d(TAG,"[VIEW LIFECYCLE EVENT] propertyChanged: " + key + ' ' + oldValue + ' ' + newValue);
 	}
+	
 }